Fixed the opacity IPF resetting to 0 if the value given was 100% or greater (fixes #2185)
This also restores the 1.12 behavior of bumping the opacity values of semi-transparent pixels
if the opacity is past 100%.

+++ src/image_modifications.cpp
@@ -541,7 +541,7 @@ surface o_modification::operator()(const surface& src) const
 		return nullptr;
 	}
 
-	Uint8 amount = ftofxp(opacity_);
+	Uint16 amount = ftofxp(opacity_);
 
 	{
 		surface_lock lock(nsurf);
